* Implemented parsing equation part
* Using regex it's now possible to parse passed equation and fullfil
with retrieved values vector with Unknown class instances.

+++ Unknown.hpp
@@ -4,15 +4,15 @@
 class Unknown
 {
 	public:
-		Unknown(size_t polynomial_degree, int number) : polynomial_degree_(polynomial_degree), number_(number) {};
+		Unknown(size_t polynomial_degree, double number) : polynomial_degree_(polynomial_degree), number_(number) {};
 		~Unknown() {}
-		void	SetNumber(int number) { number_ = number; }
-		int		GetNumber() const { return (number_); }
+		void	SetNumber(double number) { number_ = number; }
+		double		GetNumber() const { return (number_); }
 		size_t	GetPolynDegree() const { return (polynomial_degree_); }
 
 	private:
 		size_t	polynomial_degree_;
-		int		number_;
+		double		number_;
 };
 
 #endif
